Implement `Duration::round` and some general updates/fixes (#24)
* General Duration updates and cleanup

* Initial impl of methods complete

* General work completed - minus test bug

* Finish work, add tests

* Add tests for ceil and expand

* cargo clippy

* Apply review and some API cleanup

+++ src/components/duration/normalized.rs
@@ -38,9 +38,9 @@ impl NormalizedTimeDuration {
 
     // TODO: Implement as `ops::Div`
     /// `Divide the NormalizedTimeDuraiton` by a divisor.
-    pub(super) fn divide(&self, divisor: i64) -> i64 {
+    pub(super) fn divide(&self, divisor: i64) -> f64 {
         // TODO: Validate.
-        self.0 as i64 / divisor
+        self.0 / (divisor as f64)
     }
 
     /// Equivalent: 7.5.31 NormalizedTimeDurationSign ( d )
@@ -68,7 +68,7 @@ impl NormalizedTimeDuration {
 
     /// Round the current `NormalizedTimeDuration`.
     pub(super) fn round(&self, increment: u64, mode: TemporalRoundingMode) -> TemporalResult<Self> {
-        let rounded = utils::round_number_to_increment(self.0 as i64, increment, mode);
+        let rounded = utils::round_number_to_increment(self.0, increment as f64, mode);
         if rounded.abs() > MAX_TIME_DURATION as i64 {
             return Err(TemporalError::range()
                 .with_message("normalizedTimeDuration exceeds maxTimeDuration."));
@@ -93,6 +93,7 @@ impl Add<Self> for NormalizedTimeDuration {
 }
 
 /// A normalized `DurationRecord` that contains a `DateDuration` and `NormalizedTimeDuration`.
+#[derive(Debug, Clone, Copy)]
 pub struct NormalizedDurationRecord(pub(super) (DateDuration, NormalizedTimeDuration));
 
 impl NormalizedDurationRecord {
